chore: Move updateModerationStatusFromSourceStatus to DocumentTrait.php
