+++ html/modules/custom/reliefweb_entities/src/OpportunityDocumentTrait.php
@@ -2,8 +2,6 @@
 
 namespace Drupal\reliefweb_entities;
 
-use Drupal\Core\Entity\RevisionLogInterface;
-use Drupal\reliefweb_entities\Entity\Source;
 use Drupal\reliefweb_moderation\Helpers\UserPostingRightsHelper;
 use Drupal\reliefweb_utility\Helpers\TaxonomyHelper;
 use Drupal\reliefweb_utility\Helpers\UserHelper;
@@ -100,44 +98,6 @@ trait OpportunityDocumentTrait {
   protected function updateModerationStatusFromExpirationDate() {
     if ($this->getModerationStatus() === 'published' && $this->hasExpired()) {
       $this->setModerationStatus('expired');
-    }
-  }
-
-  /**
-   * Update the status to refused if any of the sources is blocked.
-   */
-  protected function updateModerationStatusFromSourceStatus() {
-    if (!$this->hasField('field_source') || $this->field_source->isEmpty()) {
-      return;
-    }
-
-    $blocked = [];
-    foreach ($this->field_source as $item) {
-      $source = $item->entity;
-      if (empty($source) || !($source instanceof Source)) {
-        continue;
-      }
-
-      if ($source->getModerationStatus() === 'blocked') {
-        $blocked[] = $source->label();
-      }
-    }
-
-    if (!empty($blocked)) {
-      $this->setModerationStatus('refused');
-
-      // Add a message to the revision log.
-      if ($this instanceof RevisionLogInterface) {
-        $message = 'Submissions from "' . implode('", "', $blocked) . '" are no longer allowed.';
-
-        $log = $this->getRevisionLogMessage();
-        if (empty($log)) {
-          $this->setRevisionLogMessage($message);
-        }
-        else {
-          $this->setRevisionLogMessage($message . ' ' . $log);
-        }
-      }
     }
   }
